requirements.txt
requests~=2.28.1
behave~=1.2.7.dev2
mysql-connector-python~=8.0.31
selenium~=4.5.0